weather_eats_agent/pyproject.toml
[project]
name = "weather-eats-agent"
version = "0.1.0"
description = "AI-powered restaurant discovery agent using live weather and Google Maps, built with Google ADK and MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.27.4",
    "python-dotenv>=1.0.0",
    "google-cloud-logging>=3.0.0",
    "anyio>=4.0.0",
]
